change credit and debit

+++ SOMEI/SOMEI/Model/Structures/PostingValueStruct.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct PostingValueStruct:Codable {
     
-    var valor:Double
+    var valor:Int
     var descricao:String
     var fixa:Bool
     var dtVencimento:DtVencimentoStruct
